Spread ChunkDimensionInfo into all of the sparse extractors.
Consolidated the sparse utilities.
